refactoring/TagProvider: extract functions, fix bug: detect duplicates with standard customs/tags on add new tag
